Add files via upload
Added landing approach sounds: 200, 100, 50, 40, 30, 20, 10. Alerts will only play when the landing gear is down to avoid unnecessary warnings. It will also only function if the aircraft supports a radio altimeter; otherwise, it will not be used to avoid confusion.

Added a caution sound alert when one of the engines stops working.